AutoMapperProfile refactor - Beautified AutoMapperProfile - Removed redundant value resolvers - Renamed 'Helpers' folder to 'Mapping'
